scripts/pyproject.toml
[project]
name = "scripts"
version = "0"
description = "Local build utilities"
dependencies = [
    "python-benedict[toml]",
    "tomli-w",
    "packaging",
]

[tool.uv]
managed = true
skip-build = true

[tool.uv_build]
package = false

